requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
psycopg2-binary
gunicorn

# pip install -r requirements.txt